Address Copilot review: break the core<->machine cycle, drop dead guard
- Shared domain types (RadioState, RadioDeps, FeedbackSound) and the
  timing constants move to radioMachine.ts — their natural owner — and
  radioCore re-exports them, so every other module keeps its single
  import surface. The dependency is now one-directional
  (adapter -> machine); no more runtime value imports going both ways.
- Remove the unused recentUserPauseIntent guard (its logic already
  lives inline in unexpectedOfflinePause).
